requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.36.0
openai>=1.40.0       
langchain>=0.2.11
langchain-openai>=0.1.21
langgraph>=0.2.13
langchain-community>=0.2.11
langchain-text-splitters>=0.2.2
faiss-cpu>=1.8.0 
tiktoken>=0.7.0
pymupdf>=1.24.4 
rapidocr-onnxruntime>=1.3.22
pandas>=2.2.2  
